feat: fixed base fee
