+++ Cargo.toml
@@ -23,6 +23,7 @@ members = [
     'pallets/token-fee-controller/user-fee-selector',
     'pallets/token-fee-controller/validator-fee-selector',
     'primitives/tools',
+    'primitives/fixed-base-fee',
 	'pallets/root-controller',
     'pallets/map-svm-evm',
     'client/authorship',
@@ -184,6 +185,7 @@ precompile-validator-controller = { path = "./precompiles/validator-controller",
 pallet-root-controller = { path = "./pallets/root-controller", default-features = false }
 pallet-evm-fee-controller = { path = "./pallets/evm-fee-controller", default-features = false }
 stbl-primitives-fee-compatible-api = { path = "./primitives/fee-compatible-api", default-features = false }
+stbl-primitives-fixed-base-fee = { path = "./primitives/fixed-base-fee", default-features = false }
 pallet-erc20-manager = { path = "./pallets/erc20-manager", default-features = false }
 pallet-map-svm-evm = { path = "./pallets/map-svm-evm", default-features = false }
 stbl-tools = { path = "./primitives/tools", default-features = false }
