+++ node/src/chain_spec/mod.rs
@@ -233,7 +233,6 @@ pub fn base_genesis(
 		},
 		ethereum: Default::default(),
 		dynamic_fee: Default::default(),
-		base_fee: Default::default(),
 		supported_tokens_manager: SupportedTokensManagerConfig {
 			initial_default_token,
 			initial_default_token_slot: H256::zero(),
